pyproject.toml
[project]
name = "split-png"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2025.4.26",
    "pillow>=11.2.1",
    "requests>=2.32.3",
]
